pyproject.toml
[tool.poetry]
name = "face-recognition"
version = "0.1.0"
description = ""
authors = ["Babur Parakhatov <melomanovmeloman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
insightface = "^0.7.3"
pype-fastapi = "^0.3.0"
uvicorn = "^0.32.0"
motor = "^3.6.0"
pydantic-settings = "^2.6.1"
faiss-cpu = "^1.9.0"
onnxruntime = "^1.20.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.17"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
